fix(ui): use component render instead of usecallback to render button
